==> 1. IMDB Data Cleaning.ipynb <==
import os
import pandas as pd

script_dir = os.getcwd() 

print(f"Current working directory: {script_dir}")

Originally, the title.basics.tsv.gz file contained the following fields: tconst, titleType, primaryTitle, originalTitle, isAdult, startYear, endYear, runtimeMinutes, and genres.

Since the original title, start year, end year, and runtime are not relevant features, I have removed them. Additionally, I have filtered out any content that is not classified as a movie (excluding "short" films).

# Load data
basics_input_file = os.path.join(script_dir, "IMDB Datasets", "title.basics.tsv")
df = pd.read_csv(basics_input_file, sep="\t", dtype=str, na_values="\\N")

# Filter irrelevant columns
title_basics_df_filtered = df.loc[df['titleType'] == 'movie', ["tconst",  "isAdult", "startYear", "runtimeMinutes", "genres"]]

display(title_basics_df_filtered.head(10))

To ensure data quality, I have also removed movies with fewer than 1,000 ratings, eliminating lesser-known films that could introduce noise into the recommendations.

# Load data
ratings_input_file = os.path.join(script_dir, "IMDB Datasets", "title.ratings.tsv") 
df = pd.read_csv(ratings_input_file, sep="\t", dtype=str, na_values="\\N")

df["numVotes"] = df["numVotes"].astype(int)

# Calculate Mean, Median, Mode
mean_votes = df["numVotes"].mean()
median_votes = df["numVotes"].median()
mode_votes = df["numVotes"].mode()[0]

# Print results
print(f"Mean: {mean_votes}")
print(f"Median: {median_votes}")
print(f"Mode: {mode_votes}")

ratings_df_filtered = df.loc[df["numVotes"] >= 1000]

display(ratings_df_filtered.head(10))

To create a comprehensive dataset, I merged the filtered movie data with ratings and crew information using the common identifier tconst. Additionally, I removed the titleType and numVotes columns, as they are no longer needed.

# Load data
crew_input_file = os.path.join(script_dir, "IMDB Datasets", "title.crew.tsv")
df_crew = pd.read_csv(crew_input_file, sep="\t", dtype=str, na_values="\\N")

# Merge dataframes
df_merged = pd.merge(title_basics_df_filtered, ratings_df_filtered, on="tconst", how="inner")
df_merged = pd.merge(df_merged, df_crew, on="tconst", how="left")

# Remove numVotes
df_merged = df_merged.drop(columns=["numVotes"])

# Save data
output_file = os.path.join(script_dir, "Cleaned Datasets", "merged_movie_data.tsv")
df_merged.to_csv(output_file, sep="\t", index=False)

# Display
print(f"Merged dataset saved to: {output_file}")
display(df_merged.head(10))
